feat: DB와 Member관련 Entity들 동기화
PK id값 LONG -> INTEGER로 일괄 변경
Auth UserName은 PK 값으로 통일
DB - Entity 동기화

issue: #6

+++ src/main/java/keeper/project/homepage/repository/MemberRepository.java
@@ -5,7 +5,7 @@ import keeper.project.homepage.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
+public interface MemberRepository extends JpaRepository<MemberEntity, Integer> {
 
   Optional<MemberEntity> findByLoginId(String loginId);
